refactor: UserSearchFormation.tsx
