+++ apps/fcdb/src/features/user-search/ui/UserSearchFormationHalfCoat.tsx
@@ -1,0 +1,112 @@
+import Image from "next/image";
+import { POSITION_LOCATIONS } from "@/shared/constant/position";
+import { useQuery } from "@tanstack/react-query";
+import { MetaQueries } from "@/entities/meta/model/queries";
+import SafeImage from "@/shared/components/SafeImage";
+import { getPositionColor, findPositionCategory } from "@/shared/lib/position";
+
+export const UserSearchFormationHalfCoat = ({
+  formation,
+  formationGroup,
+}: {
+  formation: {
+    [key: string]: {
+      spId: number;
+      spPosition: string;
+      spGrade: number;
+    };
+  };
+  formationGroup: "first" | "second";
+}) => {
+
+  const { data: soccerPlayerMeta } = useQuery(MetaQueries.getPlayerMeta());
+  const { data: seasonIdMeta } = useQuery(MetaQueries.getSeasonIdMeta());
+
+  const positionsEntries = Object.entries(POSITION_LOCATIONS);
+  const positionsList =
+    formationGroup === "first"
+      ? positionsEntries
+      : [...positionsEntries].reverse();
+
+  return (
+    // 확인 후 pb-[24px] 제거 필요
+    <div className="w-1/2 h-full grid grid-cols-9 grid-rows-5 gap-1 px-[12px] pb-[24px]">
+      {positionsList.map(([key, positions], columnIndex) =>
+        positions.map((position, rowIndex) => {
+          const soccerPlayer = formation?.[position];
+
+          if (!soccerPlayer) {
+            return null;
+          }
+
+          const getPlayerImageSrc = () => {
+            const imgSrc = `https://fco.dn.nexoncdn.co.kr/live/externalAssets/common/playersAction/p${soccerPlayer.spId}.png`;
+
+            return imgSrc;
+          };
+
+          const playerName = soccerPlayerMeta?.find(
+            (player) => player.id === soccerPlayer.spId
+          )?.name;
+          const grade = soccerPlayer.spGrade;
+          const seasonId = Number(soccerPlayer.spId.toString().slice(0, 3));
+          const seasonImg = seasonIdMeta?.find(
+            (element) => element.seasonId == seasonId
+          )?.seasonImg;
+          const positionCategory = findPositionCategory(
+            soccerPlayer.spPosition
+          );
+          const positionColor = getPositionColor(positionCategory);
+
+          return (
+            <div
+              key={position}
+              className={`w-full h-full flex flex-col items-center justify-center`}
+              style={{
+                gridColumn: columnIndex + 1,
+                gridRow: rowIndex + 1,
+              }}
+            >
+              <div className="relative w-[50px] h-[60px] flex items-center justify-center">
+                <SafeImage
+                  src={getPlayerImageSrc()}
+                  alt={playerName ? playerName : "선수 이미지"}
+                  width={50}
+                  height={60}
+                  className="rounded-full border-[1px] border-[#ABEE02]"
+                />
+                {seasonImg && (
+                  <div className="absolute left-0 -bottom-[20px] w-full">
+                    <div className="flex justify-between w-full items-center">
+                      <Image
+                        src={seasonImg}
+                        alt="season-image"
+                        width={26}
+                        height={20}
+                        className="rounded-[4px]"
+                      />
+                      <div className="border-[1px] border-[#ABEE02] h-[20px] w-[20px] flex items-center justify-center bg-[#212121] rounded-[4px]">
+                        <p className="text-[#CE535D] font-semibold text-[10px]">
+                          {grade}
+                        </p>
+                      </div>
+                    </div>
+                    <div className="w-[calc(100%+8px)] -mx-1 overflow-hidden">
+                      <p
+                        className="font-semibold text-[10px] text-center whitespace-nowrap overflow-hidden text-clip"
+                        style={{ color: positionColor }}
+                      >
+                        {playerName}
+                      </p>
+                    </div>
+                  </div>
+                )}
+              </div>
+              <div className="text-[10px] text-center" title={playerName}></div>
+            </div>
+          );
+        })
+      )}
+    </div>
+  );
+};
